+++ source/Public/New-SqlDscRole.ps1
@@ -70,6 +70,14 @@ function New-SqlDscRole
         $Refresh
     )
 
+    begin
+    {
+        if ($Force.IsPresent -and -not $Confirm)
+        {
+            $ConfirmPreference = 'None'
+        }
+    }
+
     process
     {
         if ($Refresh.IsPresent)
@@ -88,18 +96,18 @@ function New-SqlDscRole
             $PSCmdlet.ThrowTerminatingError(
                 [System.Management.Automation.ErrorRecord]::new(
                     [System.InvalidOperationException]::new($errorMessage),
-                    'NSR0001', # cspell: disable-line
+                    'NSDR0001', # cspell: disable-line
                     [System.Management.Automation.ErrorCategory]::ResourceExists,
                     $Name
                 )
             )
         }
 
-        $verboseDescriptionMessage = $script:localizedData.Role_Create_ShouldProcessVerboseDescription -f $Name, $ServerObject.InstanceName
-        $verboseWarningMessage = $script:localizedData.Role_Create_ShouldProcessVerboseWarning -f $Name
+        $descriptionMessage = $script:localizedData.Role_Create_ShouldProcessDescription -f $Name, $ServerObject.InstanceName
+        $confirmationMessage = $script:localizedData.Role_Create_ShouldProcessConfirmation -f $Name
         $captionMessage = $script:localizedData.Role_Create_ShouldProcessCaption
 
-        if ($Force.IsPresent -or $PSCmdlet.ShouldProcess($verboseDescriptionMessage, $verboseWarningMessage, $captionMessage))
+        if ($PSCmdlet.ShouldProcess($descriptionMessage, $confirmationMessage, $captionMessage))
         {
             try
             {
@@ -110,11 +118,7 @@ function New-SqlDscRole
                     $serverRole.Owner = $Owner
                 }
 
-                Write-Verbose -Message ($script:localizedData.Role_Creating -f $Name)
-
                 $serverRole.Create()
-
-                Write-Verbose -Message ($script:localizedData.Role_Created -f $Name)
 
                 return $serverRole
             }
@@ -125,7 +129,7 @@ function New-SqlDscRole
                 $PSCmdlet.ThrowTerminatingError(
                     [System.Management.Automation.ErrorRecord]::new(
                         [System.InvalidOperationException]::new($errorMessage, $_.Exception),
-                        'NSR0002', # cspell: disable-line
+                        'NSDR0002', # cspell: disable-line
                         [System.Management.Automation.ErrorCategory]::InvalidOperation,
                         $Name
                     )